finished?

+++ getUrls.py
@@ -1,11 +1,15 @@
 # UPA Project 3
-# Author: Vojtech Fiala <xfiala61>
+# Author: Vojtech Fiala <xfiala61@stud.fit.vutbr.cz>
+#
+# Support group of co-authors: 
+# Vojtech Giesl <xgiesl00@stud.fit.vutbr.cz>
+# Vojtech Kronika <xkroni01@stud.fit.vutbr.cz>
 
 import os
 from bs4 import BeautifulSoup as bs
-import requests
+import cloudscraper
 
-
+# Main class to download URLs of products
 class URLDownloader():
     def __init__(self, product):
         # careful with this, apparently CPUs are hosted on "proshop.dk" and GPUs on .de etc
@@ -20,11 +24,10 @@ class URLDownloader():
 
         self.__category = self.__base_url + '/' + self.__product
         self.__file = 'urls.txt'
+        self.__scraper = cloudscraper.create_scraper(delay=10, browser='chrome')
     
     def __getPage(self, page):
-        # headers to act like a browser on my machine
-        headers = {'user-agent':'Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/106.0.0.0 Safari/537.36'}
-        page = requests.get(page, allow_redirects=True, headers=headers)
+        page = self.__scraper.get(page)
         return page.text
 
     def __soupify(self, document):
@@ -69,7 +72,8 @@ class URLDownloader():
         self.__saveUrls(file=self.__file, urls=product_links)
         
 if __name__ == '__main__':
-    os.remove('urls.txt')   # remove file each time so that the links are fresh
+    if os.path.isfile('urls.txt'):
+        os.remove('urls.txt')   # remove file each time so that the links are fresh
     x = URLDownloader(product='CPU')
     x.getLinks()
     y = URLDownloader(product='GPU')
